Reorg + typo fix, add AtomCloudDonut to optics init
